Moved cloning to InstanceDataService
