+++ collect_app/src/main/java/org/odk/collect/android/formhierarchy/FormHierarchyFragmentHostActivity.kt
@@ -9,7 +9,6 @@ import org.odk.collect.android.formentry.FormOpeningMode
 import org.odk.collect.android.formentry.FormSessionRepository
 import org.odk.collect.android.formentry.repeats.DeleteRepeatDialogFragment
 import org.odk.collect.android.injection.DaggerUtils
-import org.odk.collect.android.instancemanagement.InstanceCloner
 import org.odk.collect.android.instancemanagement.InstancesDataService
 import org.odk.collect.android.instancemanagement.autosend.AutoSendSettingsProvider
 import org.odk.collect.android.projects.ProjectsDataService
@@ -83,9 +82,6 @@ class FormHierarchyFragmentHostActivity : LocalizedActivity() {
     @Inject
     lateinit var changeLockProvider: ChangeLockProvider
 
-    @Inject
-    lateinit var instanceCloner: InstanceCloner
-
     private val sessionId by lazy { intent.getStringExtra(EXTRA_SESSION_ID)!! }
     private val viewModelFactory by lazy {
         FormEntryViewModelFactory(
@@ -124,7 +120,7 @@ class FormHierarchyFragmentHostActivity : LocalizedActivity() {
                     viewModelFactory,
                     this,
                     scheduler,
-                    instanceCloner,
+                    instancesDataService,
                     projectsDataService.getCurrentProject().value?.uuid
                 )
             }
