+++ collect_app/src/main/java/org/odk/collect/android/formhierarchy/FormHierarchyViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import org.javarosa.core.model.FormIndex
 import org.javarosa.core.model.instance.TreeReference
-import org.odk.collect.android.instancemanagement.InstanceCloner
+import org.odk.collect.android.instancemanagement.InstancesDataService
 import org.odk.collect.android.javarosawrapper.FormController
 import org.odk.collect.androidshared.async.TrackableWorker
 import org.odk.collect.androidshared.data.Consumable
@@ -14,7 +14,8 @@ import org.odk.collect.async.Scheduler
 
 class FormHierarchyViewModel(
     scheduler: Scheduler,
-    private val instanceCloner: InstanceCloner
+    private val projectId: String,
+    private val instancesDataService: InstancesDataService
 ) : ViewModel() {
     private val trackableWorker = TrackableWorker(scheduler)
     val isEditing: LiveData<Boolean> = trackableWorker.isWorking
@@ -33,7 +34,7 @@ class FormHierarchyViewModel(
 
         trackableWorker.immediate(
             background = {
-                instanceCloner.clone(formController)
+                instancesDataService.clone(formController, projectId)
             },
             foreground = { dbId ->
                 if (dbId != null) {
@@ -47,10 +48,11 @@ class FormHierarchyViewModel(
 
     class Factory(
         private val scheduler: Scheduler,
-        private val instanceCloner: InstanceCloner
+        private val projectId: String,
+        private val instancesDataService: InstancesDataService
     ) : ViewModelProvider.Factory {
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return FormHierarchyViewModel(scheduler, instanceCloner) as T
+            return FormHierarchyViewModel(scheduler, projectId, instancesDataService) as T
         }
     }
 }
