+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -51,7 +51,6 @@ import org.odk.collect.android.formmanagement.CollectFormEntryControllerFactory;
 import org.odk.collect.android.formmanagement.OpenRosaClientProvider;
 import org.odk.collect.android.formmanagement.FormsDataService;
 import org.odk.collect.android.formmanagement.ServerFormsDetailsFetcher;
-import org.odk.collect.android.instancemanagement.InstanceCloner;
 import org.odk.collect.android.geo.MapConfiguratorProvider;
 import org.odk.collect.android.geo.MapFragmentFactoryImpl;
 import org.odk.collect.android.instancemanagement.InstancesDataService;
@@ -657,19 +656,6 @@ public class AppDependencyModule {
                 mdmConfigHandler,
                 broadcastReceiverRegister,
                 restrictionsManager
-        );
-    }
-
-    @Provides
-    public InstanceCloner providesInstanceCloner(
-            ProjectsDataService projectsDataService,
-            InstancesRepositoryProvider instancesRepositoryProvider,
-            StoragePathProvider storagePathProvider
-    ) {
-        String currentProjectId = projectsDataService.getCurrentProject().getValue().getUuid();
-        return new InstanceCloner(
-                instancesRepositoryProvider.create(currentProjectId),
-                storagePathProvider.create(currentProjectId).getInstancesDir()
         );
     }
 }
